\name{cb.city-package}
\alias{cb.city-package}
\alias{cb.city}
\docType{package}
\title{
\packageTitle{cb.city}
}
\description{
\packageDescription{cb.city}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cb.city}
\packageIndices{cb.city}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cb.city}

Maintainer: \packageMaintainer{cb.city}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
