% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cb.city.R
\docType{package}
\name{cb.city}
\alias{cb.city}
\alias{cb.city-package}
\title{Cynthia Brewer's palettes at cpt-city}
\description{
Cynthia Brewer's palettes at cpt-city
}
\examples{
list_cpt()
}
\author{
A.J. Barbour
}
\references{
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/}
}
\seealso{
\code{\link{list_cpt}} for a way to see all possible options (in \code{\link{cbcity}});
\code{\link{cpt}} and \code{\link{read_cptcity}} are responsible
for downloading the actual palettes, and \code{\link{get_cpt}} is a convenience function.
}

