% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpt.R
\name{city_palette}
\alias{city_palette}
\title{Convenience function to quickly get a color palette}
\usage{
city_palette(name, number, ...)
}
\arguments{
\item{name}{character; the palette name (case sensitives)}

\item{number}{character or integer; the palette number; this depends on \code{name}}

\item{...}{additional arguments to \code{\link{read_cpt}}}
}
\value{
A \code{\link{cpt.cols-class}} object
}
\description{
Return a color palette which can be used for plotting,
with a single command
}
\details{
This function uses \code{\link{read_cpt}} and \code{\link{cpt}}.
}
\examples{
\dontrun{
# Say we wanted to get a sequential palette of oranges to reds
Pal <- 'OrRd'
n <- 5

# The most direct way is:
p.a <- city_palette(Pal, n)

# careful, though, this isnt a vector
str(p.a)
# It can be coerced to one though:
as.character(p.a)
# or
as.vector(p.a)

# The 'long' way to get the palette:
p.b <- read_cpt(cpt(Pal, n))

# ... should yield the same result as the short way:
try(identical(p.a, p.b))

# Take a look:
layout(matrix(1:2))
plot(p.a)
plot(p.b)

}
}
\seealso{
\code{\link{read_cpt}}
}

