% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.R
\name{cpt-class}
\alias{cpt-class}
\alias{print.cpt}
\title{Methods for the 'cpt' class}
\usage{
\method{print}{cpt}(x, ...)
}
\arguments{
\item{x}{an object of class 'cpt'}

\item{...}{additional parameters; for \code{print} these are fed to \code{\link{cat}}}
}
\description{
Methods for the 'cpt' class
}
\seealso{
\code{\link{cpt}}
}

