% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{cpt.cols-class}
\alias{as.character.cpt.cols}
\alias{as.data.frame.cpt.cols}
\alias{as.vector.cpt.cols}
\alias{cpt.cols}
\alias{cpt.cols-class}
\alias{data.frame.cpt.cols}
\alias{length.cpt.cols}
\alias{plot.cpt.cols}
\alias{print.cpt.cols}
\title{Methods for the 'cpt.cols' class}
\usage{
\method{as.character}{cpt.cols}(x, ...)

\method{as.vector}{cpt.cols}(x, ...)

\method{length}{cpt.cols}(x)

\method{data.frame}{cpt.cols}(x, ...)

\method{as.data.frame}{cpt.cols}(x, ...)

\method{print}{cpt.cols}(x, ...)

\method{plot}{cpt.cols}(x, n.total = length(x), ax = NULL, ylab = NULL,
  ...)
}
\arguments{
\item{x}{an object of class 'cpt.cols'}

\item{...}{additional parameters; for \code{print} these are fed to \code{\link{cat}};
for \code{plot} these are fed to \code{\link{image}}}

\item{n.total}{integer; currently unused}

\item{ax}{statements which draw axes for the colors.
This overrides the default axes which are simply an index of color numbers
using \code{\link{axis}}}

\item{ylab}{character or expression; if this is \code{NULL} the name of
the palette is shown; could be used to design a custom color palette, for example.}
}
\description{
Methods for the 'cpt.cols' class
}
\seealso{
\code{\link{read_cpt}}
}

