% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt.cols}
\alias{as.character.cpt.cols}
\alias{as.data.frame.cpt.cols}
\alias{as.vector.cpt.cols}
\alias{cpt.cols}
\alias{data.frame.cpt.cols}
\alias{length.cpt.cols}
\alias{plot.cpt.cols}
\alias{print.cpt.cols}
\title{Methods for the 'cpt.cols' class}
\usage{
\method{as.character}{cpt.cols}(x, ...)

as.vector.cpt.cols(x, ...)

\method{length}{cpt.cols}(x)

data.frame.cpt.cols(x, ...)

\method{as.data.frame}{cpt.cols}(x, ...)

\method{print}{cpt.cols}(x, ...)

\method{plot}{cpt.cols}(x, n.total = length(x), ax = NULL, ...)
}
\arguments{
\item{x}{an object with class 'cpt.cols'}

\item{n.total}{integer; currently unused}

\item{ax}{statements which draw \code{\link{axes}} on the color axis.
This overrides the default axes which are simply an index of color numbers}

\item{x}{an object of class \code{\link{cpt.cols}}}
}
\description{
Methods for the 'cpt.cols' class
}

