% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cptcity.R
\name{cpt}
\alias{.check_cpt_list}
\alias{cpt}
\alias{print.cpt}
\title{Create a cpt object}
\usage{
cpt(name, number, check.url = TRUE)

\method{print}{cpt}(x, ...)

.check_cpt_list(name, number, error = TRUE)
}
\arguments{
\item{name}{character; the palette name (case sensitive)}

\item{number}{character or integer; the palette number}

\item{check.url}{logical; should the status of cpt's url be tested?}

\item{error}{logical; If \code{TRUE}, an error is signalled; if \code{FALSE}, only a warning is given.}
}
\value{
An object with class \code{'cpt'} with metadata for the palette
}
\description{
Create a cpt object
}
\details{
Setting \code{check.url=TRUE} can add a bit
of time, since cpt-city servers have relatively long response times, but
doing so can save time downstream.
}
\examples{
\dontrun{
cpt('OrRd',5)

# Uses an internal check to make sure palettes do exist
try(cpt('OrangeRed',5)) #fails because the palette name is wrong
try(cpt('OrRd',999)) #fails becayse there isnt an OrRd variant with that many colors
}
}
\seealso{
\code{\link{read_cptcity}} or \code{\link{get_cpt}} to download the colors, and \code{\link{list_cpt}}
to see all available options for \code{name} and \code{number}
}

