% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpt.R
\name{list_cpt}
\alias{list_cpt}
\title{List all possible color palettes}
\source{
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/}
}
\usage{
list_cpt(...)
}
\arguments{
\item{...}{additional parameters to \code{\link{subset}}}
}
\description{
List all possible color palettes
}
\details{
List all possible color palettes available on cpt-city since the last package update
}
\examples{
head(list_cpt())
list_cpt(pal.n=="8")
list_cpt(pal.name=="Set1")

with(list_cpt(pal.type=='div'), table(pal.name))
with(list_cpt(pal.type=='seq'), table(pal.name))
with(list_cpt(pal.type=='qual'), table(pal.name))
}
\seealso{
\code{\link{cbcity}} for the current list, and \code{\link{read_cpt}} or \code{\link{get_cpt}} to get a specific palette
}

