% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpt.R
\name{read_cpt}
\alias{read_cpt}
\alias{read_cpt.cpt}
\title{Download and read-in a cpt-city color palette}
\source{
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/notes/formats.html}
}
\usage{
read_cpt(x, ...)

\method{read_cpt}{cpt}(x, ...)
}
\arguments{
\item{x}{an object to read in}

\item{...}{additional parameters to \code{\link{read.table}}}
}
\value{
A \code{\link{cpt.cols-class}} object
}
\description{
Download and read-in a cpt-city color palette
}
\details{
Uses \code{\link{read.table}} to download and read in the color palette.
}
\examples{
\dontrun{

c.5 <- read_cpt(cpt('OrRd',5))
c.7 <- read_cpt(cpt('OrRd',7))

layout(matrix(1:3))
plot(c.5)
plot(c.7)
# specify the axes
plot(c.7, ax=axis(1, at=1:7, labels=letters[1:7]))
}
}
\seealso{
\code{\link{cpt}}, and \code{\link{city_palette}}, a convenience function used to download the palette in one step
}

