% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cb.city.R
\docType{package}
\name{cb.city}
\alias{cb.city}
\alias{cb.city-package}
\title{Cynthia Brewer's palettes at cpt-city}
\description{
This package provides accress to Cynthia Brewer's sequential, divergent, and qualitative palettes found online
at cpt-city [1].  It is important to choose a palette appropriate for the data, and
guidance as to which palette to use can be found at \url{http://colorbrewer2.org/}.
}
\details{
The main functions to use are \code{\link{cpt}} and \code{\link{read_cpt}}, but there is also
a convenience function \code{\link{city_palette}} to get a palette in one step.

You can use \code{\link{list_cpt}} to see all possible palettes (references built-in table \code{\link{cbcity}}).
}
\examples{
\dontrun{
head(list_cpt())
plot(get_cpt('OrRd',8))
}
}
\author{
A.J. Barbour
}
\references{
[1]: \url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/}
}
\seealso{
\code{\code{\link{list_cpt}}, \link{cpt}}, \code{\link{read_cpt}}, and \code{\link{city_palette}}
}

