% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cb.city.R
\docType{package}
\name{cb.city}
\alias{cb.city}
\alias{cb.city-package}
\title{Cynthia Brewer's palettes at cpt-city}
\description{
This package provides accress to Cynthia Brewer's sequential, divergent, and qualitative palettes online
at cpt-city [1].  For guidance as to which palette type to use, refer to [2].
}
\details{
The main functions to use are \code{\link{read_cpt}} and \code{\link{cpt}}.
You can use \code{\link{list_cpt}} to see all possible palettes (references built-in table \code{\link{cbcity}}).
}
\examples{
\dontrun{
head(list_cpt())
plot(get_cpt('OrRd',8))
}
}
\author{
A.J. Barbour
}
\references{
[1]: \url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/}

[2]: \url{http://colorbrewer2.org/}
}
\seealso{
\code{\link{cpt}} and \code{\link{read_cpt}} are responsible
for downloading the actual palettes, and \code{\link{get_cpt}} is a convenience function
which marries the two functions.
}

