% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cb.city.R
\docType{data}
\name{cbcity}
\alias{cbcity}
\title{A list of the current color palettes}
\format{\code{data.frame} with the palette type, the palette name, the palette number string,
and the palette number}
\source{
Parsed from links on
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/seq/index.html},
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/div/index.html}, and
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/cb/qual/index.html}
}
\description{
A list of the current color palettes
}
\examples{
data(cbcity)
summary(cbcity)
subset(cbcity, pal.name=="Set1")
}
\seealso{
\code{\link{cpt}}
}

