% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cptcity.R
\name{get_cpt}
\alias{get_cpt}
\title{Convenience function to quickly get a color palette}
\usage{
get_cpt(name, number)
}
\arguments{
\item{name}{character; the palette name (case sensitive)}

\item{number}{character or integer; the palette number}
}
\description{
Convenience function to quickly get a color palette
}
\details{
This uses \code{\link{read_cptcity}} and \code{\link{cpt}} to
return the color palette desired.
}
\examples{
\dontrun{
as.character(get_cpt('OrRd',5))
}
}

